requirements.txt
flask==3.0.0
openai==0.28.1
python-dotenv==1.0.0
Werkzeug>=3.0.0
click>=8.0.0
blinker>=1.6.0
itsdangerous>=2.0.0
Jinja2>=3.0.0
MarkupSafe>=2.0.0